Separate pybind11 translation unit for Filter
