+++ src/python_pybind11/src/_ignition_math_pybind11.cc
@@ -169,20 +169,14 @@ PYBIND11_MODULE(math, m)
 
   ignition::math::python::defineMathFrustum(m, "Frustum");
 
-  ignition::math::python::defineMathFilter<int>(m, "Filteri");
-  ignition::math::python::defineMathFilter<float>(m, "Filterf");
-  ignition::math::python::defineMathFilter<double>(m, "Filterd");
+  ignition::math::python::defineMathFilter(m, "Filter");
 
-  ignition::math::python::defineMathBiQuad<int>(m, "BiQuadi");
-  ignition::math::python::defineMathBiQuad<float>(m, "BiQuadf");
-  ignition::math::python::defineMathBiQuad<double>(m, "BiQuadd");
+  ignition::math::python::defineMathBiQuad(m, "BiQuad");
 
   ignition::math::python::defineMathBiQuadVector3(
     m, "BiQuadVector3");
 
-  ignition::math::python::defineMathOnePole<int>(m, "OnePolei");
-  ignition::math::python::defineMathOnePole<float>(m, "OnePolef");
-  ignition::math::python::defineMathOnePole<double>(m, "OnePoled");
+  ignition::math::python::defineMathOnePole(m, "OnePole");
 
   ignition::math::python::defineMathOnePoleQuaternion(
     m, "OnePoleQuaternion");
